Create Azure-WindowsVM-Setup.txt
This project documents the step-by-step process of creating a Windows 10 Virtual Machine (VM) in Microsoft Azure. It includes setting up a Resource Group, Virtual Network (VNet), and Subnet, as well as configuring security settings like Trusted Launch, Secure Boot, and vTPM.

The goal of this lab is to gain hands-on experience with Azure infrastructure, including VM provisioning, networking, and security settings. This project is useful for IT Help Desk, Cloud Security, and Cybersecurity beginners who want to understand the fundamentals of managing cloud-based virtual machines.